requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Celery and Redis
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Streamlit
streamlit==1.28.2
plotly==5.17.0
pandas==2.1.4

# Monitoring and Logging
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
croniter==2.0.1
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
